Fix raw times
The problem is that when we do a convergence study, we want to get the
raw times used by the solver, but our implementation doesn't quite allow
that right now. Also, this problem would be hard to fix properly without
breaking down all the notebooks. So instead, I add an optional
parameters that makes the solver use raw times by default.