Use hpc_backend in jacobian
Supposedly, it should be enough to only use the hpc backend when
creating arrays (e.g. y_r and the identity matrix in the case of the
jacobian), in order to avoid unnecessary memory transfers. Numpy should
be smart enough to not transfer the array when using other methods.

Actually, Cupy does not allow implicit memory transfers, so we should be
pretty safe here.