Fix partial update
I'm currently saving the full old y to calculate the partial update to F since the force interactions between only the perturbed cells need to be updated.
This can be done by subtracting the old force interactions between the perturbed cells and adding the new ones.
It might be possible to save only the relevant part of the old y with some indexing magic.